Move data scheme dgm to test_utils
