+++ p2panda-encryption/src/data_scheme/test_utils/mod.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 pub mod dcgka;
+pub mod dgm;
 pub mod network;
 pub mod ordering;
